langchain structured output/requirements.txt
huggingface_hub
langchain_huggingface
python-dotenv
langchain 
pydantic